Added inotify watcher

+++ src/main.c
@@ -4,19 +4,27 @@
 #include <string.h>
 #include <proc/procps.h>
 #include <dirent.h>
+#include <pthread.h>
 #include "SCLogger.h"
 #include "procutils.h"
 #include "configutils.h"
 #include "tcpcomm.h"
 #include "basement.h"
+#include "watcher.h"
+
+/**
+ * List of exit codes:
+ *
+ * 0  ... OK
+ * -1 ... could not delete a file
+ */
 
 int main(int argc, char **argv)
 {
-    int* processList = NULL;
-    char *cpDir = "/proc/";
     CONFIG config;
     ZOMBIE_NODE* head = NULL;
     char szLogMessage[1024] = {0};
+    pthread_t watcherId;
 
     memset(&config, 0, sizeof(CONFIG));
 
@@ -28,20 +36,39 @@ int main(int argc, char **argv)
     if(config.checkIntervall == 0)
     	config.checkIntervall = DEFAULT_CHECK_INTERVALL;
 
+    int iWatcher = initWatcher();
+    addDirectory(iWatcher, "/home/flichten");
+    pthread_create(&watcherId, NULL, watch, (void*)(intptr_t)iWatcher);
+
     while(1)
     {
+    	if(checkStopFlag(STOP_FLAG_FILE))
+    	{
+    		break;
+    	}
+
+    	/*
         processList = getProcessList(cpDir);
 
         checkProcessState(processList);
 
         processList = NULL;
+        */
 
         sleep(config.checkIntervall);
     }
 
     logIt(INFO, "End");
 
-    destroyLogging();
+    //end procedures
+	pthread_cancel(watcherId);
+
+	if(removeFile(STOP_FLAG_FILE) != 0)
+	{
+		return -1;
+	}
+
+	destroyLogging();
 
 	return 0;
 }
